refactor: move models to internal
