+++ repositories/post_reposititory.go
@@ -1,7 +1,7 @@
 package repositories
 
 import (
-	"echo-demo-project/models"
+	"echo-demo-project/internal/models"
 
 	"gorm.io/gorm"
 )
