+++ server/handlers/post_handler.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"echo-demo-project/models"
+	"echo-demo-project/internal/models"
 	"echo-demo-project/repositories"
 	"echo-demo-project/requests"
 	"echo-demo-project/responses"
