+++ services/post/create.go
@@ -1,6 +1,6 @@
 package post
 
-import "echo-demo-project/models"
+import "echo-demo-project/internal/models"
 
 func (postService *Service) Create(post *models.Post) {
 	postService.DB.Create(post)
