+++ services/post/delete.go
@@ -1,6 +1,6 @@
 package post
 
-import "echo-demo-project/models"
+import "echo-demo-project/internal/models"
 
 func (postService *Service) Delete(post *models.Post) {
 	postService.DB.Delete(post)
