+++ services/token/create_refresh_token.go
@@ -1,7 +1,7 @@
 package token
 
 import (
-	"echo-demo-project/models"
+	"echo-demo-project/internal/models"
 	"time"
 
 	"github.com/golang-jwt/jwt/v5"
