+++ services/token/service.go
@@ -2,7 +2,7 @@ package token
 
 import (
 	"echo-demo-project/config"
-	"echo-demo-project/models"
+	"echo-demo-project/internal/models"
 
 	"github.com/golang-jwt/jwt/v5"
 )
